[common] move normalization utils into it's own folder
