+++ packages/common/src/Utils/normalization/normalizeInitialState.js
