+++ packages/common/src/index.js
@@ -61,7 +61,7 @@ export { getToolbarTheme } from './Utils/getToolbarTheme';
 export { simplePubsub } from './Utils/simplePubsub';
 export { getModalStyles } from './Utils/getModalStyles';
 export { mergeStyles } from './Utils/mergeStyles';
-export { default as normalizeInitialState } from './Utils/normalizeInitialState';
+export { default as normalizeInitialState } from './Utils/normalization/normalizeInitialState';
 export { validate, getContentStateSchema } from './Utils/data-schema-validator';
 export { isHexColor } from './Components/ColorPicker/utils';
